refactor: rename sccs_layout_check to utils
